requirements.txt
streamlit>=1.28.0
python-docx>=1.1.0
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
pytesseract>=0.3.10
openai>=1.0.0
requests>=2.31.0